Refactoring SplashScreen to MVP pattern.
