+++ app/src/main/java/br/com/ecostage/mobilecollect/ui/splashscreen/SplashScreenPresenter.kt
@@ -1,0 +1,11 @@
+package br.com.ecostage.mobilecollect.ui.splashscreen
+
+/**
+ * Splash Screen presenter operations.
+ *
+ * Created by andremaia on 8/27/17.
+ */
+interface SplashScreenPresenter {
+    fun setupCrashReports()
+    fun initializeFirebase()
+}
