+++ app/src/main/java/br/com/ecostage/mobilecollect/ui/splashscreen/SplashScreenView.kt
@@ -1,0 +1,10 @@
+package br.com.ecostage.mobilecollect.ui.splashscreen
+
+/**
+ * Splash screen view operations.
+ *
+ * Created by andremaia on 8/27/17.
+ */
+interface SplashScreenView {
+    fun goToNextView()
+}
